% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annex_K.R
\name{annex_K}
\alias{annex_K}
\title{Add annex K lines at the end of modules and imp_exp flows}
\usage{
annex_K(path)
}
\arguments{
\item{path}{Path where your files are located}
}
\value{

}
\description{
Add annex K lines at the end of modules and imp_exp flows
}
\examples{
path <- "C:/Users/fr103343/Documents/R/Projects/annex 4/conversion/2020-12-10/conversion"
annex_K(path)
}
