% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_results.R
\name{read_results}
\alias{read_results}
\title{Read Team Excel results}
\usage{
read_results(path, skip_row = 1)
}
\arguments{
\item{skip_row}{}
}
\value{
dataframe
}
\description{
Read Team Excel results
}
